Don't warn if invokable test setup function doesn't exist
This was never the intention (as the code comments imply), but
it was never tested.

This is a cherry-pick of 50f234df500829a0023ed5d396c486f995ad71ef
because it went to dev (5.13) when it should have originally went
to 5.12.

Change-Id: I8df0b3702129b1f1d086df73117d3ddb721317cb
Reviewed-by: Gatis Paeglis <gatis.paeglis@qt.io>
(cherry picked from commit 50f234df500829a0023ed5d396c486f995ad71ef)
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>