Fix incorrect font size in certain styles on Windows
When a style is loaded, QtQuickControls2Plugin::loadStylePlugins() is
called. This function uses QPluginLoader to search the style directory
for the style plugin. On Windows, the code was looking for the plugin
without the debug "d" suffix, causing the style plugin to fail to load
for debug builds. This results in the QQuickTheme subclasses not being
created, causing controls to be shown with default font sizes.

This patch fixes the issue by appending the "d" suffix on Windows.

Change-Id: I706dbe39325a104bcd6ce72cb0a8d37025adb055
Fixes: QTBUG-71902
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>