QML: Consider all possible metatypes when looking for attached props
metaObjectToType is in fact a MultiHash. There can be more than one QML
type associated with a given metaobject. Therefore, in order to find the
attached properties we need to iterate them.

This highlights a different problem: If you register the same metaobject
under different QML names and define different attached properties for
each, you will get random results here. That, however, is a problem of
the public qmlAttachedPropertiesObject function, which doesn't take a
parameter for the registered name of the type.

Change-Id: I0eb9618114029ef5000d8eb12691b7f8de485121
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>