Fix TextField background not respecting control's width
After 6858d4e9, the background of the Material style TextField does not respect
the width of the control if the control is resized (e.g. due to being in a
layout).

- The first time resizeBackground() is called, !extra.isAllocated() returns
  true - that is, extra has not been allocated yet.
- The if statement is executed and the background's width is set to the
  available width of the control (e.g. 120).
- As a result of the background's width being set, its widthValid flag is set
  to true. This would previously be undone directly afterwards by setting
  widthValid flag to false.
- In the case of the test case, the implicitWidth was already 120,
  so geometryChanged is not emitted. However, when the height of the background
  is set, this *does* cause itemGeometryChanged() to be called, which in turn
  does the following:

      extra.value().hasBackgroundWidth = p->widthValid;
      extra.value().hasBackgroundHeight = p->heightValid;
      resizeBackground();

  So now all of the following checks evaluate to false:

      (!p->widthValid || !extra.isAllocated() || !extra->hasBackgroundWidth)

  This prevents the background from being resized.

This patch fixes the issue by unsetting the widthValid (and heightValid) flags
of the background directly after setting its width. To be safe, it also only
unsets it if we were the ones to set it. By doing this, the check mentioned
above succeeds because p->widthValid and extra->hasBackgroundWidth are both now
false.

It also adds some extra safety into itemGeometryChanged() that ensures that
extra.value().hasBackgroundWidth is only set if necessary, and therefore
prevents the extra data from unnecessarily being allocated. This is not
necessary for the fix, but feels like the right thing to do.

Change-Id: I051e281718bd8a2a20c100767d929fb71497ce1b
Fixes: QTBUG-71875
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>