TableModel: add QML-invokable overloads for index(), data() and setData()
data() needs to be invokable because it does more processing to ensure that
DisplayRole will give us something, and we want to be able to access it in
JS code too, not only via the role context property in a delegate binding.
index() needs to be invokable to use it when calling data().  It's useful
for setData() to be invokable so that TableView delegates can be used to
edit the model.  However since we don't normally expose numeric roles to
QML, we use string roles, and have to look them up in the roleNames() hash.

Change-Id: I38904ac995fc2bac514bde2dd37a95e0b911c00c
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>