TableModel: add roleDataProvider callback
As an alternative to trying to write smarter C++ in the data() accessor, we
give the user full control of data conversion by calling an external JS
function if defined, to map role to the value that data() should return.
This enables extracting arbitrary values, converting the data in arbitrary
ways, or even doing calculations in case the EditRole stores a formula and the
DisplayRole should provide the result, or something like that.

This callback is implemented somewhat like TableView.columnWidthProvider, but
the arguments are more complex: function(row, column, role, rawData)

Change-Id: Ifaf5807f4809e0b5ad1d1c403f65c0707b902f10
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>