Fix lookups the transition between primitive strings and string objects
Suppose we have

    function foo(x) { return x.constructor; }

and we call it first with

    foo("hello")

then the lookup will be initialized with a primitive getter from the
prototype.

When we subsequently call

    foo(new String("world"))

then the primitiveGetterProto() will check that the provided object is
of the same type as last time, which erroneously succeeds. Indeed, both
are of Managed type. However now we're passing a full-fledged object,
which is not a primitive anymore - hence the additional check to ensure
that we fall back to the generic getter.

Task-number: QTBUG-69898
Change-Id: I3c7a8384bfdf0e31b7c6247cce80fe7448b627b3
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>