Property Cache: Use related meta objects to check for "enums"
If a method of a QObject-derived class takes a "foreign" enum as a
parameter, the enum's meta object is added to the "relatedMetaObjects"
property. Therefore, we can actually determine if something is a named
enumerator, even if we initially don't have its metaobject.

This way we don't need to detect these types by their type flags anymore
and therefore we can remove the band aid added in commit
fd9d88e3a9f24eb84d029fb9324b4f03ae58ffeb.

Change-Id: Id602d2b6f688a12b5addef855f6a16920b4e6549
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>