Add basic type annotation support for QML functions to the grammar
Inspired by TypeScript syntax, allow optional type annotations in the
style of ": <name of type>" in for the parameters of functions and their
return type.

These annotations are not used at the moment, so by default we produce
an error message when encountering them in the AST.

In addition their usage is limited to functions declared in the QML
scope. All other uses attempt to produce readable syntax errors. So for
example this is okay:

    Item {
        function blah(param: string) string { ... }
    }

And this is not okay:

    // some file.js
    function blah(param: string) : string { ... }

Change-Id: I12d20b4e2ce92e01108132975a06cfd13ba4a254
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>