QQuickAbstractButton: fix clicked() not being emitted after long press
When a button is pressed long enough
(QStyleHints::mousePressAndHoldInterval()) and there is a connection
(e.g. signal handler) to the pressAndHold() signal, it is emitted.
If nothing is connected to the signal, clicked() is emitted.

Before this patch, QQuickAbstractButton used
QObjectPrivate::isSignalConnected() to check whether or not anything
was connected to pressAndHold(). The problem with this function is
described by Olivier:

   "[...] there is an optimisation for the first 64 signals
    we store a bit in a 64bit integer to quickly see if a signal is
    connected. There is also the problem that it can return true even
    if the signal has been disconnected."

This is also mentioned in a comment in the code:

    Returns \c true if the signal with index \a signal_index from object
    \a sender is connected. Signals with indices above a certain range
    are always considered connected (see connectedSignals in
    QObjectPrivate).

When 5adce04 added inset signals to QQuickControl, it meant that
signals in QQuickAbstractButton (which derives from QQuickControl) were
pushed outside of that 64 signal range, resulting in the
IS_SIGNAL_CONNECTED macro returning true.

This patch fixes the issue by using QObject::isSignalConnected(), which
does not use the 64 signal optimization.

Fixes: QTBUG-72811
Change-Id: Ic6e54d6cab062e528522ef7e3cf27c1023d31347
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>