Fix font and palette settings in .conf files not being respected
In 5.11, the font palette read from the qtquickcontrols2.conf file was
respected. After some large architectural changes (around the time of
94780538) required to fix a crash, 5.12 stopped respecting these
settings.

This patch fixes the issue by setting the default font as the System
scope font, because that's what QQuickControlPrivate::parentFont() uses
as its fallback if no parent item has a font explicitly set.
QQuickControlPrivate::parentFont() is used as the starting point for
font inheritance/resolution for each control.

The same fix is used for palettes.

Change-Id: I706a9f109c9959b8ea6b91f842146dbfc876cb2b
Fixes: QTBUG-72023
Reviewed-by: Nils Jeisecke <nils.jeisecke@saltation.com>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>