Ask rendering thread to release resources when window is closed
If the close event is accepted, the close handler in QWindow will
destroy the underlying platform window without notifying the
QSGRenderer attribute windowManager. This may cause a race
condition if the renderer is a QSGThreadedRenderLoop because
the platform window may be deleted while the thread is using it.

Correct this issue by notifying the QSGRenderer on a close event
using the hide() method (which is supposed to release the
resources according to documentation).

Fixes: QTBUG-82474
Task-number: QTBUG-73929
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit ca66d4fc23706a0f4ecfc4031f37cdd19facd836)
Change-Id: Ia8d76a4471270ca90af7c201e7ad4477609b9c36