V4: Add label for loop in spread in ArrayPattern
This patch also rotates the loop back so that the condition is at the
top of the loop. It's a cherry-picked from commit
0282b89ec672e25a465a8e51bc74c7fd58a624b1.

Without explicitly setting the label, we get a default constructed
Label from labelForOffset in PlatformAssemblerCommon::link, which leads
to a jump into nirvana.

This issue arises only with backward jumps, as we fill in the
information for forward jump targets once we actually encounter the
target.

Fixes: QTBUG-77047
Change-Id: Id928831f90eace494adb1eb1190f674a6f033b20
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>