Move propertyCache-by-minor-version into QQmlMetaTypeData
Ths data structure should only be accessed when protected by the
metatype data lock. In fact we don't access it from anywhere else.
To make that more obvious, move it to the right place.

This allows us to eliminate some const_cast and poking around in
QQmlTypePrivate from the outside.

Change-Id: I16ffd240b9504b9c00010bdb2d17b05c8196fe8a
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>