QQuickScrollView: only forward content size to flickable if set explicit
TableView monitors if the application sets an explicit
contentWidth/Height, and skips calculating it when that is the case.
So be more careful from ScrollView, and don't set it on the
flickable if the application did not set it explicitly.
By doing it this way, you can now set an explicit contentWidth/Height
both in ScrollView and TableView, and it will be respected.

Task-number: QTBUG-72536
Change-Id: Ib87fa6958881fccdc47f50133e996484392281b6
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>