Fix DialogButtonBox content size calculation
Some history:

- f1f884d3 worked around an issue in DialogButtonBox.
- c2fd8f7d fixed it by using contentWidth; i.e. the implicit width of
  the contentItem. It caused QTBUG-72372.
- I tried to fix QTBUG-72372 with 6476de0b, but created (or exposed)
  QTBUG-73860.

The problem in QTBUG-73860 can be seen with the following example:

    Dialog {
        id: dialog
        visible: true
        standardButtons: Dialog.Ok
    }

The single 'Ok' button here will go outside of the dialog. The
underlying issue can be seen by looking into DialogButtonBox.qml:

    implicitWidth: Math.max(implicitBackgroundWidth + leftInset + rightInset,
        (control.count === 1 ? contentWidth * 2 : contentWidth) + leftPadding + rightPadding)
    implicitHeight: Math.max(implicitBackgroundHeight + topInset + bottomInset,
        contentHeight + topPadding + bottomPadding)

The implicit width of the box in this case is contentWidth * 2
(there is one button, so control.count === 1). This should result in
the button taking half the width of the box and being aligned to the
right:

    alignment: count === 1 ? Qt.AlignRight : undefined

    ...

    delegate: Button {
        width: control.count === 1 ? control.availableWidth / 2 : undefined
    }

What actually happens is that the contentItem (ListView) is temporarily 0
until it gets its final size of 100. However, QQuickDialogButtonBox
doesn't respond to this change in the ListView's contentWidth.

This problem can be fixed by returning to c2fd8f7d's resizeContent()
implementation, which uses contentWidth.

Then, there is a second issue:

    Dialog {
        id: dialog
        visible: true
        standardButtons: Dialog.Ok
        width: 300
    }

The button here is also positioned outside of the box. The problem is
that the contentWidth is based on implicitContentWidth:

    QQuickContainerPrivate::updateContentWidth()
    {
        // ...
        contentWidth = implicitContentWidth;
        // ...
    }

implicitContentWidth is calculated by calling getContentWidth():

    void QQuickControlPrivate::updateImplicitContentWidth()
    {
        // ...
        implicitContentWidth = getContentWidth();
        // ...
    }

In the case of horizontal alignment,
QQuickDialogButtonBoxPrivate::getContentWidth() uses the implicit
width of the largest button:

    for (int i = 0; i < count; ++i) {
        QQuickItem *item = q->itemAt(i);
        if (item) {
            totalWidth += item->implicitWidth();
            maxWidth = qMax(maxWidth, item->implicitWidth());
        }
    }

    // ...

    if ((alignment & Qt::AlignHorizontal_Mask) == 0)
        totalWidth = qMax(totalWidth, count * maxWidth + totalSpacing);

The Default style button has an implicitWidth of 100.
The DialogButtonBox in the example above is 300 pixels wide, so the
button should be 150, and it is, thanks to its width binding.
However, the DialogButtonBox uses contentWidth to size its contentItem
(ListView), and the contentWidth is, as mentioned, 100: the implicit
width of the button. So, the button ends up hanging over the side
of the box, as it's larger than the box thinks it is.

This problem is fixed by setting DialogButtonBox's contentWidth to the
contentWidth of the contentItem (ListView). This makes DialogButtonBox
use the explicit widths of the buttons rather than their implicit
widths. Since the contentWidth is no longer implicit, we must also
change any use of contentWidth in the implicitWidth binding to
implicitContentWidth.

While writing auto tests for this, they caught an issue where
contentWidth wasn't updated, so now we call resizeContent() in
QQuickContainer::setContentWidth().

Change-Id: I99ffda21b47aeb14d4382e453e87c4312f343a1c
Fixes: QTBUG-72886
Fixes: QTBUG-73860
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>