TableModel.roleDataProvider: replace row, column with index argument
QModelIndex has various advantages over separate int row and column in
TableModel API; we intend to expose it via an invokable index() method and
via a context property for delegates, and use it in the invokable data()
function.  So we should be consistent by using it in roleDataProvider too.
This way the callback only requires 3 arguments instead of 4, and the first
one is potentially extensible, just in case.

Change-Id: I7b1bc9ea5adb64941979d83901b3566278357e98
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>