QQuickScrollView: respect the content size set on/by the flickable
If the flickable inside a scrollview has an explicit content
size assigned, or if the flickable was created by the
application, then don't fall back to calculate the content
size. The reason is that it should not try to change the content
size of a flickable that is controlled by the application (or by
the flickable itself, in the case of ListView, ScrollView and
TextArea). Sometimes e.g ListView will report a negative
contentWidth, which is not considered illegal.

From before we used to just check if the child flickable had
a negative content size, and if so, take that as a evidence
that the flickable was owned by the scrollview. But with this
patch, we instead add two extra variables that keeps explicit
track of whether or not we should read the content size from the
flickable, regardless of the values it might return.

With the two new variables, we also no longer need the
"ownsFlickable" property, as we can instead use the new
variables to check for the same.

Fixes: QTBUG-72536
Fixes: QTBUG-74000
Fixes: QTBUG-72940
Change-Id: Iec87cc953557bf8c1bdb32a3c11b721c607fc19a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>