Fix leaking scene graph rendering contexts
When using the threaded render loop, the rendering thread assumes
ownership of the rendering context. If the rendering thread is never
started, that context is leaked. This happens in tests sometimes when we
create and destroy a window without waiting for exposure.

So this patch maintains the ownership in the render loop for created
contexts until the per-window thread is started.

Task-number: QTBUG-74348
Change-Id: Ifa397fab0833c82110ac4571c1e3790351c43afd
Reviewed-by: BogDan Vatra <bogdan@kdab.com>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>